pg_stat_statements: Allow interrupts whilst reading query text file
When the query text file gets very large (hundreds of megabytes, or more)
and when limited I/O capacity is available, reading it can take many
seconds, or even minutes.

In such situations we want statement timeouts and query cancellations
to allow canceling the read half-way through. To support this, check
for interrupts, and read 100MB at a time, instead of 1GB as before.